consolidated EnvironmentVariableProvider
