+++ src/ado2gh/GithubApiFactory.cs
@@ -27,7 +27,7 @@ namespace OctoshiftCLI.AdoToGithub
         public virtual GithubApi Create(string apiUrl = null, string targetPersonalAccessToken = null)
         {
             apiUrl ??= DEFAULT_API_URL;
-            targetPersonalAccessToken ??= _environmentVariableProvider.GithubPersonalAccessToken();
+            targetPersonalAccessToken ??= _environmentVariableProvider.TargetGithubPersonalAccessToken();
             var githubClient = new GithubClient(_octoLogger, _client, _versionProvider, _retryPolicy, _dateTimeProvider, targetPersonalAccessToken);
             return new GithubApi(githubClient, apiUrl, _retryPolicy);
         }
