+++ src/gei/Handlers/MigrateRepoCommandHandler.cs
@@ -430,5 +430,5 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
 
     private string GetAzureStorageConnectionString(MigrateRepoCommandArgs args) => args.AzureStorageConnectionString.HasValue()
         ? args.AzureStorageConnectionString
-        : _environmentVariableProvider.AzureStorageConnectionString();
+        : _environmentVariableProvider.AzureStorageConnectionString(false);
 }
